+++ src/helper.ts
@@ -1,115 +1,180 @@
-"use strict";
+import { existsSync, readFileSync, statSync } from "fs";
+import { basename, dirname, join as joinPath, parse as parsePath } from "path";
+import { TextDocument, TextEditor, window, workspace } from "vscode";
 
-import * as vscode from "vscode";
-import * as path from "path";
+import { html as jsBeautify } from "js-beautify";
+import { getExtension, getType as getMimeType } from "mime";
+import mjml2html from "mjml";
 
-import fileUrl = require("file-url");
-import { mjml2html } from "mjml";
-import { existsSync, readFileSync } from "fs";
 import * as ejs from 'ejs';
 
-export default class Helper {
+export async function renderMJML(cb: (content: string) => any, fixImg?: boolean, minify?: boolean, beautify?: boolean): Promise<void> {
+  const activeTextEditor: TextEditor | undefined = window.activeTextEditor;
+  if (!activeTextEditor) {
+    return;
+  }
 
-    static async renderMJML(cb: (content: string) => any, fixLinks?: boolean, minify?: boolean, beautify?: boolean): Promise<any> {
-        if (!(this.isMJMLFile(vscode.window.activeTextEditor.document))) {
-            vscode.window.showWarningMessage("This is not a MJML document!");
-            return;
+  if (!isMJMLFile(activeTextEditor.document)) {
+    window.showWarningMessage("This is not a MJML document!");
+    return;
+  }
+
+  let content = await compileHtml(activeTextEditor.document.getText());
+
+  content = mjmlToHtml(
+    content,
+    minify !== undefined ? minify : workspace.getConfiguration("mjml").minifyHtmlOutput,
+    beautify !== undefined ? beautify : workspace.getConfiguration("mjml").beautifyHtmlOutput
+  ).html;
+
+  if (content) {
+    if (fixImg !== undefined && fixImg) {
+      content = fixImages(content, getPath());
+    }
+
+    return cb(content);
+  } else {
+    window.showErrorMessage(`MJMLError: Failed to parse file ${basename(getPath())}`);
+  }
+}
+
+export function isMJMLFile(document: TextDocument): boolean {
+  return document.languageId === "mjml" && (document.uri.scheme === "file" || document.uri.scheme === "untitled");
+}
+
+export function mjmlToHtml(
+  mjml: string, minify: boolean, beautify: boolean, path?: string, validation: "strict" | "soft" | "skip" = "skip"
+): { html: string, errors: any[] } {
+  try {
+    if (!path) {
+      path = getPath();
+    }
+
+    return mjml2html(mjml, {
+      beautify,
+      filePath: path,
+      minify,
+      mjmlConfigPath: getCWD(path),
+      validationLevel: validation
+    });
+  } catch (error) {
+    return { html: "", errors: [error] };
+  }
+}
+
+function resolveFileOrText(fileName: string): string | undefined {
+  let document = workspace.textDocuments.find(e => e.fileName === fileName);
+
+  if (document) {
+      return document.getText();
+  }
+  if (dirname(fileName) && existsSync(fileName)) {
+      return readFileSync(fileName, "utf8");
+  }
+
+  return undefined;
+}
+
+export async function compileHtml(html: string): Promise<string> {
+  const activeTextEditor: TextEditor | undefined = window.activeTextEditor;
+  if (!activeTextEditor) {
+    throw new Error(`No active editor`);
+  }
+
+  let currentFileName = activeTextEditor.document.fileName;
+  const fileFolder = currentFileName.substring(0, currentFileName.lastIndexOf("/"));
+  let dataFileName = `${fileFolder}/preview.json`;
+
+  const dataSource = resolveFileOrText(dataFileName);
+
+  if (dataSource) {
+    const config = JSON.parse(dataSource);
+    const data = config.data;
+
+    if (config.helpers) {
+        const helperFileName = `${fileFolder}/${config.helpers}`;
+        const helpers = await import(helperFileName);
+        data.helpers = helpers;
+    }
+    const compiled = ejs.compile(html, { filename: currentFileName });
+    return compiled(data);
+  }
+
+  return html;
+}
+
+export function fixImages(text: string, mjmlPath: string): string {
+    return text.replace(
+        new RegExp(/((?:src|url)(?:=|\()(?:[\'\"]|))((?!http|\\|"|#).+?)([\'\"]|\))/, "gmi"),
+        (_1: string, start: string, src: string, end: string): string => {
+            return start + encodeImage(joinPath(dirname(mjmlPath), src), src) + end;
         }
+    );
+}
 
-        let content = await this.compileHtml(vscode.window.activeTextEditor.document.getText());
-
-        content = this.mjml2html(
-            content,
-            minify != undefined ? minify : vscode.workspace.getConfiguration("mjml").minifyHtmlOutput,
-            beautify != undefined ? beautify : vscode.workspace.getConfiguration("mjml").beautifyHtmlOutput,
+export function beautifyHTML(mjml: string): string | undefined {
+    try {
+        const replaced: string = mjml.replace(
+            new RegExp(/<.*mj-style[^>]*>(?:[^<>]+)<.*\/.*mj-style>/, "gmi"), (style: string): string => {
+                return style.replace(/mj-style/gi, "style");
+            }
         );
 
-        if (content) {
-            if (fixLinks != undefined && fixLinks) {
-                content = this.fixLinks(content);
-            }
+        const beautified: string = jsBeautify(replaced, workspace.getConfiguration("mjml").beautify);
 
-            return cb(content);
+        if (replaced !== mjml) {
+            return beautified.replace(
+                new RegExp(/<.*style[^>]*>(?:[^<>]+)<.*\/.*style>/, "gmi"), (styleBlock: string): string => {
+                    return styleBlock.replace(
+                        new RegExp(/<.*style.*>/, "gi"), (style: string): string => {
+                            return style.replace("style", "mj-style");
+                        }
+                    );
+                }
+            );
         }
-        else {
-            vscode.window.showErrorMessage(`MJMLError: Failed to parse file ${path.basename(vscode.window.activeTextEditor.document.uri.fsPath)}`);
+
+        return beautified;
+    } catch (error) {
+        window.showErrorMessage(error);
+
+        return;
+    }
+}
+
+export function getPath(): string {
+  if (window.activeTextEditor && window.activeTextEditor.document) {
+      return window.activeTextEditor.document.uri.fsPath;
+  }
+
+  return "";
+}
+
+function getCWD(mjmlPath?: string): string {
+    if (workspace.rootPath) {
+        return workspace.rootPath;
+    }
+
+    return (mjmlPath) ? parsePath(mjmlPath).dir : "";
+}
+
+function encodeImage(filePath: string, original: string): string {
+    const mimeType: string | null = getMimeType(filePath);
+    if (!mimeType) {
+        return original;
+    }
+
+    const extension: string | null = getExtension(mimeType);
+    if (!extension || ["bmp", "gif", "jpeg", "jpg", "png", "svg"].indexOf(extension) === -1) {
+        return original;
+    }
+
+    if (filePath && existsSync(filePath) && statSync(filePath).isFile()) {
+        const data: Buffer = readFileSync(filePath);
+        if (data) {
+            return `data:${mimeType};base64,${data.toString("base64")}`;
         }
     }
 
-    static isMJMLFile(document: vscode.TextDocument): boolean {
-        return document.languageId === "mjml" && document.uri.scheme !== "mjml-preview";
-    }
-
-    static mjml2html(mjml: string, minify: boolean, beautify: boolean): string {
-        try {
-            let html: any = mjml2html(mjml, {
-                level: "skip",
-                minify: minify,
-                beautify: beautify,
-                filePath: vscode.window.activeTextEditor.document.uri.fsPath,
-                cwd: this.getCWD()
-            });
-
-            if (html.html) {
-                return html.html;
-            }
-        }
-        catch (e) {
-            return;
-        }
-    }
-
-    static async compileHtml(html: string): Promise<string> {
-        let currentFileName = vscode.window.activeTextEditor.document.fileName;
-        const fileFolder = currentFileName.substring(0, currentFileName.lastIndexOf("/"));
-        let dataFileName = `${fileFolder}/preview.json`;
-
-        const dataSource = this.resolveFileOrText(dataFileName);
-
-        if (dataSource) {
-            const config = JSON.parse(dataSource);
-            const data = config.data;
-
-            if (config.helpers) {
-                const helperFileName = `${fileFolder}/${config.helpers}`;
-                const helpers = await import(helperFileName);
-                data.helpers = helpers;
-            }
-            const compiled = ejs.compile(html, { filename: currentFileName });
-            return compiled(data);
-        }
-
-        return html;
-    }
-
-    static getCWD(): string {
-        if (vscode.workspace.rootPath) {
-            return vscode.workspace.rootPath;
-        }
-        else {
-            return path.parse(vscode.window.activeTextEditor.document.uri.fsPath).dir;
-        }
-    }
-
-    static fixLinks(text: string): string {
-        return text.replace(
-            new RegExp(/((?:src|href|url)(?:=|\()(?:[\'\"]|))((?!http|\\|"|#).+?)([\'\"]|\))/, "gmi"),
-            (subString: string, p1: string, p2: string, p3: string): string => {
-                return [p1, fileUrl(path.join(path.dirname(vscode.window.activeTextEditor.document.uri.fsPath), p2)), p3].join("");
-            }
-        );
-    }
-
-    static resolveFileOrText(fileName: string): string | undefined {
-        let document = vscode.workspace.textDocuments.find(e => e.fileName === fileName);
-
-        if (document) {
-            return document.getText();
-        }
-        if (path.dirname(fileName) && existsSync(fileName)) {
-            return readFileSync(fileName, "utf8");
-        }
-
-        return undefined;
-    }
+    return original;
 }